增加基于时间判断的数据库限制

+++ miniprogram/util/database/comment.js
@@ -39,6 +39,9 @@ export class Comment {
    * @param {object} comment 需要包含 reply,text,images
    */
   async addComment(superId, superType, comment) {
+    console.warn('TODO:调用处没有修改')
+    if (!db.perControl.limitTimeStrategy('addComment', 3000))
+      return db.perControl.refusePromise()
     let commentId = util.randomId()
     if (comment.constructor != Object) {
       console.error("comment类型非法")
@@ -65,6 +68,9 @@ export class Comment {
    * @param {string} commentId 
    */
   async removeComment(commentId) {
+    console.warn('TODO:调用处没有修改')
+    if (!db.perControl.limitTimeStrategy('removeComment', 1000))
+      return db.perControl.refusePromise()
     await _db.collection('comment').doc(commentId).get().then(res => {
       wx.cloud.deleteFile({
         fileList: res.data.images
